Refactor dataset tools
