+++ osmo_jupyter/dataset/__init__.py
@@ -1,0 +1,17 @@
+# Re-export functions for backwards compatibility
+from .combine import (  # noqa: F401
+    open_and_combine_picolog_and_calibration_data,
+    get_equilibration_boundaries,
+    pivot_process_experiment_results_on_ROI,
+    open_and_combine_process_experiment_results,
+    get_all_experiment_images,
+    filter_equilibrated_images,
+    open_and_combine_and_filter_source_data,
+)
+
+from .parse import (  # noqa: F401
+    parse_ysi_kordss_file,
+    parse_ysi_classic_file,
+    parse_picolog_file,
+    parse_calibration_log_file,
+)
